Fixes the container __call methody.
The __call method now calls the parent DI Container __call method
in the event that the requested item exists in the container.
If it does not exist then it runs the router code we had before.
Unfortantly we still can't use __callStatic, that would be the
most elegant solution.

This also resolves #1.
codixor you can now grab the request object from with in the route.

+++ src/Router.php
@@ -304,11 +304,18 @@ class Router extends Container
 	 */
 	public function __call($name, $args)
 	{
-		if (empty(self::$router))
+		if ($this->offsetExists($name))
 		{
-			throw new RuntimeException('You need to install a router first!');
+			return parent::__call($name, $args);
 		}
+		else
+		{
+			if (empty(self::$router))
+			{
+				throw new RuntimeException('You need to install a router first!');
+			}
 
-		return call_user_func_array([self::$router, $name], $args);
+			return call_user_func_array([self::$router, $name], $args);
+		}
 	}
 }
